+++ lib/types.ts
@@ -217,6 +217,11 @@ export type UploadResult = {
   isValid: boolean;
   key: string;
 };
+
 export interface BearerTokenPayload {
   formID?: string;
+}
+export interface VaultFormResponses {
+  Items: { [key: string]: unknown | undefined } | Record<string, unknown> | unknown;
+  Count: number;
 }